# Optimize quantization kernels