--- a/nncf/extensions/include/common_defs.h
+++ b/nncf/extensions/include/common_defs.h
@@ -1,0 +1,9 @@
+#ifndef _COMMON_DEFS_H_
+#define _COMMON_DEFS_H_
+
+
+#define CHECK_CPU(x) TORCH_CHECK(!x.is_cuda(), #x " must be a CPU tensor")
+#define CHECK_CUDA(x) AT_ASSERTM(x.is_cuda(), #x " must be a CUDA tensor")
+#define CHECK_CONTIGUOUS(x) AT_ASSERTM(x.is_contiguous(), #x " must be contiguous")
+
+#endif // _COMMON_DEFS_H_