--- a/nncf/extensions/src/binarization/cuda/functions_cuda_impl.cu
+++ b/nncf/extensions/src/binarization/cuda/functions_cuda_impl.cu
@@ -1,159 +1,7 @@
-#include <ATen/ATen.h>
-
-#include <cuda.h>
-#include <cuda_runtime.h>
-
-#include <vector>
-#include <THC/THC.h>
-
-const int CUDA_WARP_SIZE = 32;
-const int CUDA_GRID_SIZE = 56; // #SM*2
-const int CUDA_NUM_THREADS = 1024;
-
-inline int GET_BLOCKS(const int N) {
-    return (N + CUDA_NUM_THREADS - 1) / CUDA_NUM_THREADS;
-}
+#include "common_cuda_defs.cuh"
+#include "common_cuda_funcs.cuh"
 
 namespace {
-
-__device__ bool last_block(int* counter) {
-    __threadfence();
-
-    int last = 0;
-    if (threadIdx.x == 0) {
-        last = atomicAdd(counter, 1);
-    }
-
-    return __syncthreads_or(last == gridDim.x - 1);
-}
-// support only warp size = 32
-template <typename scalar_t>
-__device__ void sum_warp(volatile scalar_t* sharr) {
-    int tidx = threadIdx.x & 31;
-    if (tidx < 16) {
-        sharr[tidx] += sharr[tidx + 16];
-        sharr[tidx] += sharr[tidx + 8];
-        sharr[tidx] += sharr[tidx + 4];
-        sharr[tidx] += sharr[tidx + 2];
-        sharr[tidx] += sharr[tidx + 1];
-    }
-}
-
-// Since volatile c10::Half arithmetic is not supported, will have to sacrifice
-// the implicit warp-synchronous programming in favor of explicit intra-warp thread
-// synchronization
-
-template <typename scalar_t>
-__device__ void sum_warp_with_explicit_sync(scalar_t* sharr) {
-    int tidx = threadIdx.x & 31;
-    if (tidx < 16) {
-        sharr[tidx] += sharr[tidx + 16];
-    }
-    __syncwarp();
-    if (tidx < 16) {
-        sharr[tidx] += sharr[tidx + 8];
-    }
-    __syncwarp();
-    if (tidx < 16) {
-        sharr[tidx] += sharr[tidx + 4];
-    }
-    __syncwarp();
-    if (tidx < 16) {
-        sharr[tidx] += sharr[tidx + 2];
-    }
-    __syncwarp();
-    if (tidx < 16) {
-        sharr[tidx] += sharr[tidx + 1];
-    }
-    __syncwarp();
-}
-
-template <typename scalar_t>
-__device__ inline void gather_warp_execution_results(scalar_t* sharr, const int tidx) {
-    sharr[tidx] = tidx * CUDA_WARP_SIZE < CUDA_NUM_THREADS ? sharr[tidx * CUDA_WARP_SIZE] : static_cast<scalar_t>(0.0);
-}
-
-template <typename scalar_t>
-__device__ void sumReduce(
-        scalar_t* __restrict__ sh_grad,
-        scalar_t sum,
-        const int tidx,
-        const int bidx,
-        scalar_t* __restrict__ dev_tmp,
-        int* __restrict__ dev_last_block_counter,
-        scalar_t* __restrict__ grad) {
-    sh_grad[tidx] = sum;
-
-    __syncthreads();
-    sum_warp(sh_grad + (tidx & ~(CUDA_WARP_SIZE - 1)));
-
-    __syncthreads();
-    if (tidx < CUDA_WARP_SIZE) {
-        gather_warp_execution_results(sh_grad, tidx);
-        sum_warp(sh_grad);
-        if (tidx == 0) {
-            dev_tmp[bidx] = sh_grad[0];
-        }
-    }
-
-    if (last_block(dev_last_block_counter)) {
-        sh_grad[tidx] = tidx < gridDim.x ? dev_tmp[tidx] : static_cast<scalar_t>(0.0);
-
-        __syncthreads();
-        sum_warp(sh_grad + (tidx & ~(CUDA_WARP_SIZE - 1)));
-
-        __syncthreads();
-        if (tidx < CUDA_WARP_SIZE) {
-            gather_warp_execution_results(sh_grad, tidx);
-            sum_warp(sh_grad);
-            if (tidx == 0) {
-                grad[0] = sh_grad[0];
-            }
-        }
-    }
-}
-
-
-template<>
-__device__ void sumReduce<c10::Half>(
-        c10::Half* __restrict__ sh_grad,
-        c10::Half sum,
-        const int tidx,
-        const int bidx,
-        c10::Half* __restrict__ dev_tmp,
-        int* __restrict__ dev_last_block_counter,
-        c10::Half* __restrict__ grad) {
-    sh_grad[tidx] = sum;
-
-    __syncthreads();
-    sum_warp_with_explicit_sync(sh_grad + (tidx & ~(CUDA_WARP_SIZE - 1)));
-
-    __syncthreads();
-    if (tidx < CUDA_WARP_SIZE) {
-        gather_warp_execution_results(sh_grad, tidx);
-        sum_warp_with_explicit_sync(sh_grad);
-        if (tidx == 0) {
-            dev_tmp[bidx] = sh_grad[0];
-        }
-    }
-
-    if (last_block(dev_last_block_counter)) {
-        sh_grad[tidx] = tidx < gridDim.x ? dev_tmp[tidx] : static_cast<c10::Half>(0.0);
-
-        __syncthreads();
-        sum_warp_with_explicit_sync(sh_grad + (tidx & ~(CUDA_WARP_SIZE - 1)));
-
-        __syncthreads();
-        if (tidx < CUDA_WARP_SIZE) {
-            gather_warp_execution_results(sh_grad, tidx);
-            sum_warp_with_explicit_sync(sh_grad);
-            if (tidx == 0) {
-                grad[0] = sh_grad[0];
-            }
-        }
-    }
-}
-
 
 template <typename scalar_t>
 __global__ void wb_cuda_scale_calc_kernel(
@@ -162,10 +10,10 @@ __global__ void wb_cuda_scale_calc_kernel(
         scalar_t* __restrict__ dev_tmp,
         int* __restrict__ dev_last_block_counter,
         const int64_t total_elements_count) {
-    const int tidx = threadIdx.x;
-    const int bidx = blockIdx.x;
-    const int gtidx = bidx * CUDA_NUM_THREADS + tidx;
-    const int grid_size = CUDA_NUM_THREADS * gridDim.x;
+    const uint16_t tidx = threadIdx.x;
+    const uint32_t bidx = blockIdx.x;
+    const uint64_t gtidx = bidx * CUDA_MAX_NUM_THREADS_PER_BLOCK + tidx;
+    const uint64_t grid_size = CUDA_MAX_NUM_THREADS_PER_BLOCK * gridDim.x;
 
     scalar_t sum = 0;
     for (int i = gtidx; i < total_elements_count; i += grid_size) {
@@ -174,8 +22,8 @@ __global__ void wb_cuda_scale_calc_kernel(
 
     sum /= total_elements_count;
 
-    __shared__ scalar_t sh_mem[CUDA_NUM_THREADS];
-    sumReduce<scalar_t>(sh_mem, sum, tidx, bidx, dev_tmp, dev_last_block_counter, scale_output);
+    __shared__ scalar_t sh_mem[CUDA_MAX_NUM_THREADS_PER_BLOCK];
+    reduce_with_shared_memory<scalar_t>(sh_mem, sum, tidx, bidx, dev_tmp, dev_last_block_counter, scale_output, gridDim.x);
 }
 
 template <typename scalar_t>
@@ -240,10 +88,10 @@ __global__ void ab_cuda_grad_scale_kernel(
         scalar_t* __restrict__ dev_tmp,
         int* __restrict__ dev_last_block_counter,
         const int64_t total_elements_count) {
-    const int tidx = threadIdx.x;
-    const int bidx = blockIdx.x;
-    const int gtidx = bidx * CUDA_NUM_THREADS + tidx;
-    const int grid_size = CUDA_NUM_THREADS * gridDim.x;
+    const uint16_t tidx = threadIdx.x;
+    const uint32_t bidx = blockIdx.x;
+    const uint32_t gtidx = bidx * CUDA_MAX_NUM_THREADS_PER_BLOCK + tidx;
+    const uint64_t grid_size = CUDA_MAX_NUM_THREADS_PER_BLOCK * gridDim.x;
 
     scalar_t sum = 0;
     for (int i = gtidx; i < total_elements_count; i += grid_size) {
@@ -252,8 +100,8 @@ __global__ void ab_cuda_grad_scale_kernel(
         sum += (*(input + i) < *scale) ? err_element * grad_element : grad_element;
     }
 
-    __shared__ scalar_t sh_mem[CUDA_NUM_THREADS];
-    sumReduce<scalar_t>(sh_mem, sum, tidx, bidx, dev_tmp, dev_last_block_counter, grad_scale);
+    __shared__ scalar_t sh_mem[CUDA_MAX_NUM_THREADS_PER_BLOCK];
+    reduce_with_shared_memory<scalar_t>(sh_mem, sum, tidx, bidx, dev_tmp, dev_last_block_counter, grad_scale, gridDim.x);
 }
 
 template <typename scalar_t>
@@ -268,10 +116,10 @@ __global__ void ab_cuda_grad_thresholds_kernel(
         int64_t contiguous_elements_per_threshold,
         int64_t threshold_count,
         int64_t channel_offset) {
-    const int tidx = threadIdx.x;
-    const int bidx = blockIdx.x;
-    const int gtidx = bidx * CUDA_NUM_THREADS + tidx;
-    const int grid_size = CUDA_NUM_THREADS * gridDim.x;
+    const uint16_t tidx = threadIdx.x;
+    const uint32_t bidx = blockIdx.x;
+    const uint32_t gtidx = bidx * CUDA_MAX_NUM_THREADS_PER_BLOCK + tidx;
+    const uint64_t grid_size = CUDA_MAX_NUM_THREADS_PER_BLOCK * gridDim.x;
 
     scalar_t sum = 0;
     for (int i = gtidx; i < total_elements_per_threshold; i += grid_size) {
@@ -290,8 +138,8 @@ __global__ void ab_cuda_grad_thresholds_kernel(
         }
     }
 
-    __shared__ scalar_t sh_mem[CUDA_NUM_THREADS];
-    sumReduce<scalar_t>(sh_mem, sum, tidx, bidx, dev_tmp, dev_last_block_counter, grad_thresholds);
+    __shared__ scalar_t sh_mem[CUDA_MAX_NUM_THREADS_PER_BLOCK];
+    reduce_with_shared_memory<scalar_t>(sh_mem, sum, tidx, bidx, dev_tmp, dev_last_block_counter, grad_thresholds, gridDim.x);
 }
 
 }
@@ -308,7 +156,7 @@ at::Tensor wb_cuda_forward(
     auto scale = at::zeros({scale_count}, input.options());
     elements_per_scale = input_elements_count / input.size(0);
 
-    auto grid_size = std::min(GET_BLOCKS(elements_per_scale), CUDA_GRID_SIZE);
+    auto grid_size = std::min(GET_BLOCKS(elements_per_scale), CUDA_BLOCKS_PER_GRID_FOR_UNIFORM_ELTWISE);
     auto dev_tmp = at::empty({grid_size}, input.options());
     auto dev_last_block_counter = at::zeros({1},  at::device(input.options().device()).dtype(at::kInt));
 
@@ -318,7 +166,7 @@ at::Tensor wb_cuda_forward(
     for (int ch_idx = 0; ch_idx < scale_count; ch_idx++)
     {
         AT_DISPATCH_FLOATING_TYPES_AND_HALF(input.type(), "wb_cuda_forward_scale", ([&] {
-          wb_cuda_scale_calc_kernel<scalar_t><<<grid_size, CUDA_NUM_THREADS, 0, at::cuda::getCurrentCUDAStream()>>>(
+          wb_cuda_scale_calc_kernel<scalar_t><<<grid_size, CUDA_MAX_NUM_THREADS_PER_BLOCK, 0, at::cuda::getCurrentCUDAStream()>>>(
               input.data<scalar_t>() + ch_idx * elements_per_scale,
               scale.data<scalar_t>() + ch_idx,
               dev_tmp.data<scalar_t>(),
@@ -330,7 +178,7 @@ at::Tensor wb_cuda_forward(
     }
 
     AT_DISPATCH_FLOATING_TYPES_AND_HALF(input.type(), "wb_cuda_forward_binarize", ([&] {
-      wb_cuda_binarize_kernel<scalar_t><<<GET_BLOCKS(input_elements_count), CUDA_NUM_THREADS, 0, at::cuda::getCurrentCUDAStream()>>>(
+      wb_cuda_binarize_kernel<scalar_t><<<GET_BLOCKS(input_elements_count), CUDA_MAX_NUM_THREADS_PER_BLOCK, 0, at::cuda::getCurrentCUDAStream()>>>(
           output.data<scalar_t>(),
           input.data<scalar_t>(),
           scale.data<scalar_t>(),
@@ -358,7 +206,7 @@ at::Tensor ab_cuda_forward(
     auto output = at::empty_like(input);
 
     AT_DISPATCH_FLOATING_TYPES_AND_HALF(input.type(), "ab_cuda_forward", ([&] {
-      ab_cuda_forward_kernel<scalar_t><<<GET_BLOCKS(input_elements_count), CUDA_NUM_THREADS, 0, at::cuda::getCurrentCUDAStream()>>>(
+      ab_cuda_forward_kernel<scalar_t><<<GET_BLOCKS(input_elements_count), CUDA_MAX_NUM_THREADS_PER_BLOCK, 0, at::cuda::getCurrentCUDAStream()>>>(
           output.data<scalar_t>(),
           input.data<scalar_t>(),
           scale.data<scalar_t>(),
@@ -403,7 +251,7 @@ std::vector<at::Tensor> ab_cuda_backward(
     int64_t contiguous_elements_per_threshold = input_elements_count / input.size(0) / input.size(1);
 
     AT_DISPATCH_FLOATING_TYPES_AND_HALF(input.type(), "ab_cuda_backward", ([&] {
-      ab_cuda_grad_input_kernel<scalar_t><<<GET_BLOCKS(input_elements_count), CUDA_NUM_THREADS, 0, at::cuda::getCurrentCUDAStream()>>>(
+      ab_cuda_grad_input_kernel<scalar_t><<<GET_BLOCKS(input_elements_count), CUDA_MAX_NUM_THREADS_PER_BLOCK, 0, at::cuda::getCurrentCUDAStream()>>>(
           grad_input.data<scalar_t>(),
           grad_output.data<scalar_t>(),
           input.data<scalar_t>(),
@@ -412,13 +260,13 @@ std::vector<at::Tensor> ab_cuda_backward(
           );
     }));
 
-    auto grid_size = std::min(GET_BLOCKS(input_elements_count), CUDA_GRID_SIZE);
+    auto grid_size = std::min(GET_BLOCKS(input_elements_count), CUDA_BLOCKS_PER_GRID_FOR_UNIFORM_ELTWISE);
     auto dev_tmp = at::empty({grid_size}, grad_output.options());
     auto dev_last_block_counter = at::zeros({1},  at::device(grad_output.options().device()).dtype(at::kInt));
 
 
     AT_DISPATCH_FLOATING_TYPES_AND_HALF(input.type(), "ab_cuda_backward", ([&] {
-          ab_cuda_grad_scale_kernel<scalar_t><<<grid_size, CUDA_NUM_THREADS, 0, at::cuda::getCurrentCUDAStream()>>>(
+          ab_cuda_grad_scale_kernel<scalar_t><<<grid_size, CUDA_MAX_NUM_THREADS_PER_BLOCK, 0, at::cuda::getCurrentCUDAStream()>>>(
               grad_scale.data<scalar_t>(),
               grad_output.data<scalar_t>(),
               output.data<scalar_t>(),
@@ -429,7 +277,7 @@ std::vector<at::Tensor> ab_cuda_backward(
               input_elements_count);
         }));
 
-    grid_size = std::min(GET_BLOCKS(total_elements_per_threshold), CUDA_GRID_SIZE);
+    grid_size = std::min(GET_BLOCKS(total_elements_per_threshold), CUDA_BLOCKS_PER_GRID_FOR_UNIFORM_ELTWISE);
     dev_tmp = at::empty({grid_size}, grad_output.options());
     dev_last_block_counter = at::zeros({1},  at::device(grad_output.options().device()).dtype(at::kInt));
 
@@ -438,7 +286,7 @@ std::vector<at::Tensor> ab_cuda_backward(
     {
         auto init_element_offset = contiguous_elements_per_threshold * ch_idx;
         AT_DISPATCH_FLOATING_TYPES_AND_HALF(input.type(), "ab_cuda_backward", ([&] {
-          ab_cuda_grad_thresholds_kernel<scalar_t><<<grid_size, CUDA_NUM_THREADS, 0, at::cuda::getCurrentCUDAStream()>>>(
+          ab_cuda_grad_thresholds_kernel<scalar_t><<<grid_size, CUDA_MAX_NUM_THREADS_PER_BLOCK, 0, at::cuda::getCurrentCUDAStream()>>>(
               grad_thresholds.data<scalar_t>() + ch_idx,
               grad_output.data<scalar_t>() + init_element_offset,
               input.data<scalar_t>() + init_element_offset,