--- a/setup.py
+++ b/setup.py
@@ -125,14 +125,6 @@ EXTRAS_REQUIRE = {
     "docs": []
 }
 
-package_data = {'nncf': ['quantization/cpu/functions_cpu.cpp',
-                         'quantization/cuda/functions_cuda.cpp',
-                         'quantization/cuda/functions_cuda_kernel.cu',
-                         'binarization/cpu/functions_cpu.cpp',
-                         'binarization/cuda/functions_cuda.cpp',
-                         'binarization/cuda/functions_cuda_kernel.cu']}
-
-
 setuptools.setup(
     name="nncf",
     version=find_version(os.path.join(here, "nncf/version.py")),
@@ -151,7 +143,6 @@ setuptools.setup(
     ],
     install_requires=INSTALL_REQUIRES,
     extras_require=EXTRAS_REQUIRE,
-    package_data=package_data,
     keywords=KEY
 )