Switched to using 2D grid for per-channel kernels and cleaned up
the code duplications across the extensions.