@vshampor 2020-07-08
Internal benchmark results:

```
CUDA low batch
------------------------------------------------
Custom Symmetric (cuda 0 ) impl:
input size: [2, 96, 112, 112]
Before:
Forward: min 0.195ms / avg 0.350ms | Backward: min 0.232ms / avg 0.561ms
After:
Forward: min 0.196ms / avg 0.354ms | Backward: min 0.235ms / avg 0.535ms

Custom Symmetric Per Weight Channel  (cuda 0 ) impl
input size: [2, 96, 112, 112]
Before:
Forward: min 0.197ms / avg 0.379ms | Backward: min 0.249ms / avg 0.650ms
After:
Forward: min 0.196ms / avg 0.356ms | Backward: min 0.249ms / avg 0.556ms

Custom Symmetric Per Activation Channel  (cuda 0 ) impl
input size: [2, 96, 112, 112]
Before:
Forward: min 0.198ms / avg 0.417ms | Backward: min 1.004ms / avg 1.690ms
After:
Forward: min 0.197ms / avg 0.407ms | Backward: min 0.266ms / avg 0.692ms

Custom Asymmetric (cuda 0 ) impl:
input size: [2, 96, 112, 112]
Before:
Forward: min 0.618ms / avg 0.862ms | Backward: min 0.254ms / avg 0.595ms
After:
Forward: min 0.624ms / avg 0.864ms | Backward: min 0.239ms / avg 0.564ms

Custom Asymmetric Per Weight Channel  (cuda 0 ) impl
input size: [2, 96, 112, 112]
Before:
Forward: min 0.633ms / avg 0.905ms | Backward: min 0.268ms / avg 0.640ms
After:
Forward: min 0.634ms / avg 0.901ms | Backward: min 0.249ms / avg 0.618ms

Custom Asymmetric Per Activation Channel  (cuda 0 ) impl
input size: [2, 96, 112, 112]
Before:
Forward: min 0.623ms / avg 1.571ms | Backward: min 1.008ms / avg 1.919ms
After:
Forward: min 0.633ms / avg 0.897ms | Backward: min 0.278ms / avg 0.576ms

CUDA high batch
------------------------------------------------
Custom Symmetric (cuda 0 ) impl:
input size: [128, 96, 112, 112]
Before:
Forward: min 3.755ms / avg 3.804ms | Backward: min 6.175ms / avg 6.269ms
After:
Forward: min 3.782ms / avg 3.864ms | Backward: min 6.187ms / avg 6.288ms


Custom Symmetric Per Weight Channel  (cuda 0 ) impl
input size: [128, 96, 112, 112]
Before:
Forward: min 3.787ms / avg 3.801ms | Backward: min 5.844ms / avg 5.913ms
After:
Forward: min 3.779ms / avg 3.848ms | Backward: min 5.955ms / avg 6.052ms



Custom Symmetric Per Activation Channel  (cuda 0 ) impl
input size: [128, 96, 112, 112]
Before:
Forward: min 3.801ms / avg 3.846ms | Backward: min 6.083ms / avg 6.194ms
After:
Forward: min 3.783ms / avg 3.836ms | Backward: min 6.521ms / avg 6.604ms

Custom Asymmetric (cuda 0 ) impl:
input size: [128, 96, 112, 112]
Before:
Forward: min 4.198ms / avg 4.297ms | Backward: min 6.135ms / avg 6.288ms
After:
Forward: min 4.238ms / avg 4.385ms | Backward: min 6.127ms / avg 6.274ms


Custom Asymmetric Per Weight Channel  (cuda 0 ) impl
input size: [128, 96, 112, 112]
Before:
Forward: min 4.206ms / avg 4.317ms | Backward: min 5.938ms / avg 6.086ms
After:
Forward: min 4.241ms / avg 4.392ms | Backward: min 5.890ms / avg 6.016ms

Custom Asymmetric Per Activation Channel  (cuda 0 ) impl
input size: [128, 96, 112, 112]
Before:
Forward: min 4.227ms / avg 4.282ms | Backward: min 6.107ms / avg 6.183ms
After:
Forward: min 4.244ms / avg 4.356ms | Backward: min 6.545ms / avg 6.636ms
```

While at high batches the improvements are not quite visible, for lower batches the speedup for the per-activation kernel version is drastic.